% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Data generation for extODAL}
\usage{
data_generator(N, beta, K, n)
}
\arguments{
\item{N}{total number of observations}

\item{beta}{p*1 vector (coefficients of the covariates)}

\item{K}{total number of sites}

\item{n}{number of patients in each site}
}
\value{
data
}
\description{
Data generation for extODAL
}
