% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator_hetero.R
\name{data_generator_hetero}
\alias{data_generator_hetero}
\title{Heterogeneous data generation for extODAL}
\usage{
data_generator_hetero(N, beta, K, n)
}
\arguments{
\item{N}{total number of observations}

\item{beta}{p*1 vector (coefficients of the covariates)}

\item{K}{total number of sites}

\item{n}{number of patients in each site}
}
\value{
data
}
\description{
Heterogeneous data generation for extODAL
}
