% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL_hetero.R
\name{ODAL_hetero}
\alias{ODAL_hetero}
\title{extension of ODAL -- heterogeneity}
\usage{
ODAL_hetero(Data, beta_true)
}
\arguments{
\item{Data}{heterogeneous data}

\item{beta_true}{true value of coefficients}
}
\value{
Bias of pooled, ODAL, and local estimators
}
\description{
extension of ODAL -- heterogeneity
}
