% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAL_homo.R
\name{ODAL_homo}
\alias{ODAL_homo}
\title{replication of ODAL}
\usage{
ODAL_homo(Data, beta_true)
}
\arguments{
\item{Data}{homogeneous data}

\item{beta_true}{true value of coefficients}
}
\value{
mean MSE of pooled, ODAL, and local estimators
}
\description{
replication of ODAL
}
