% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extODAL_hetero.R
\name{extODAL_hetero}
\alias{extODAL_hetero}
\title{Run extODAL simulation -- homogeneous (i.e., replication of ODAL)}
\usage{
extODAL_hetero(Nsim, setting, parallel_run = FALSE, plotit = FALSE)
}
\arguments{
\item{Nsim}{an integer, total number of iterations}

\item{setting}{setting of the extended ODAL simulation, ("1" or "2")}

\item{parallel_run}{if the simulation run in parallel (default is "FALSE")}

\item{plotit}{if a plot will be made (default is "FALSE")}
}
\value{
MSE of three methods (Pooled, local, and ODAL)
}
\description{
Run extODAL simulation -- homogeneous (i.e., replication of ODAL)
}
