% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extODAL_homo.R
\name{extODAL_homo}
\alias{extODAL_homo}
\title{Run extODAL simulation -- homogeneous (i.e., replication of ODAL)}
\usage{
extODAL_homo(Nsim, setting, parallel_run = FALSE, plotit = TRUE)
}
\arguments{
\item{Nsim}{an integer, total number of iterations}

\item{setting}{setting of the simulation, ("A" or "B")}

\item{parallel_run}{if the simulation run in parallel (default is "FALSE")}

\item{plotit}{if a plot will be made (default is "TRUE")}
}
\value{
mean MSEs of covariates for three methods (Pooled, local, and ODAL)
}
\description{
Run extODAL simulation -- homogeneous (i.e., replication of ODAL)
}
